backend/requirements.txt
uvicorn 
fastapi
httpx
pydantic
python-multipart
asyncpg
pypdf
python-docx
